requirements.txt
git+https://github.com/Rapptz/discord.py
discord~=1.0.1
pypresence~=4.0.0
python-dotenv~=0.15.0